Resume_scanner/requirements.txt
streamlit
scikit-learn
pdfplumber
python-docx